requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
arpeggio==2.0.2
    # via parver
attrs==23.2.0
    # via parver
black==23.12.1
    # via -r requirements.in
click==8.1.7
    # via black
dill==0.3.7
    # via pulumi
frgray-network @ git+https://github.com/frgray/pulumi-aws-py-network.git@v0.1.0
    # via -r requirements.in
grpcio==1.56.2
    # via pulumi
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via black
parver==0.5
    # via pulumi-aws
pathspec==0.12.1
    # via black
platformdirs==4.1.0
    # via black
protobuf==4.25.2
    # via pulumi
pulumi==3.103.1
    # via
    #   -r requirements.in
    #   frgray-network
    #   pulumi-aws
pulumi-aws==6.18.2
    # via
    #   -r requirements.in
    #   frgray-network
pyyaml==6.0.1
    # via pulumi
semver==2.13.0
    # via
    #   pulumi
    #   pulumi-aws
six==1.16.0
    # via pulumi
